requirement.txt
tensorflow==2.12.0
pandas==1.5.3
numpy==1.24.2
scikit-learn==1.2.1
tifffile==2023.2.28
pyyaml==6.0
tqdm==4.64.1
